requirements.txt
PyYAML~=6.0.2
jinja2~=3.1.5